+++ docs/about/history.rst
@@ -13,6 +13,13 @@ Release Notes
 |version|
 ---------
 
+Changed
+^^^^^^^
+.. bullet_list::
+
+   -  Add *Render Texture Graphics Format* option to *Animated Waves Sim Settings* to solve precision issues when using height inputs.
+
+
 Fixed
 ^^^^^
 .. bullet_list::